chore(ds): several refactoring on datagrid
